+++ packages/core-data/src/components/SearchResultsTable.js
@@ -1,17 +1,8 @@
 // @flow
 
 import React from 'react';
-import {
-  useInstantSearch,
-  usePagination
-} from 'react-instantsearch';
-import {
-  ChevronLeftIcon,
-  ChevronRightIcon
-} from '@radix-ui/react-icons';
-import clsx from 'clsx';
-import RowsPerPage from './RowsPerPage';
-import i18n from '../i18n/i18n';
+import { useInstantSearch } from 'react-instantsearch';
+import Pagination from './Pagination';
 
 type Column = {
   /**
@@ -36,72 +27,6 @@ type Props = {
    */
   columns: Array<Column>
 }
-
-const Pagination = () => {
-  const refinePagination = usePagination().refine;
-  const { results } = useInstantSearch();
-
-  // const { hitsPerPage } = results;
-
-  const firstItemIndex = (results.page * results.hitsPerPage) + 1;
-  const lastItemIndex = firstItemIndex + results.hits.length - 1;
-
-  const canGoForward = (firstItemIndex + results.hitsPerPage) < results.nbHits;
-  const canGoBack = firstItemIndex > results.hitsPerPage;
-
-  const onNextPage = () => refinePagination(results.page + 1);
-  const onPreviousPage = () => refinePagination(results.page - 1);
-
-  return (
-    <div className='w-full h-10 bg-white flex justify-end items-center gap-4 text-sm px-4'>
-      <span className='font-bold'>
-        {i18n.t('SearchResultsTable.rowsPerPage')}
-      </span>
-      <RowsPerPage />
-      <span>
-        {firstItemIndex}
-        -
-        {lastItemIndex}
-        &nbsp;
-        {i18n.t('Common.words.of')}
-        &nbsp;
-        {results.nbHits}
-      </span>
-      <button
-        className={clsx(
-          'h-6',
-          'w-6',
-          'rounded-full',
-          'flex',
-          'items-center',
-          'justify-center',
-          { 'text-gray-300': !canGoBack }
-        )}
-        disabled={!canGoBack}
-        onClick={onPreviousPage}
-        type='button'
-      >
-        <ChevronLeftIcon />
-      </button>
-      <button
-        className={clsx(
-          'h-6',
-          'w-6',
-          'rounded-full',
-          'flex',
-          'items-center',
-          'justify-center',
-          { 'text-gray-300': !canGoForward }
-        )}
-        disabled={!canGoForward}
-        onClick={onNextPage}
-        type='button'
-      >
-        <ChevronRightIcon />
-      </button>
-    </div>
-  );
-};
 
 const SearchResultsTable = (props: Props) => {
   const { results } = useInstantSearch();
