+++ packages/core-data/tailwind.config.js
@@ -14,12 +14,6 @@ export default {
         gray: {
           '1000': '#505A6A'
         },
-        neutral: {
-          '100': '#F5F5F5',
-          '200': '#E5E5E5',
-          '300': '#D4D4D4',
-          '800': '#262626'
-        },
         muted: '#0005119e',
         primary: '#073B4C',
         secondary: colors.neutral['200']
